backend/pyproject.toml
[project]
name = "rastro-backend"
version = "0.2.0"
requires-python = ">=3.12"
dependencies = [
    # PRD 1
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "psycopg2-binary==2.9.10",
    "pgvector==0.3.6",
    "alembic==1.14.0",
    "pydantic==2.10.0",
    "pydantic-settings==2.6.0",
    "python-jose[cryptography]==3.3.0",
    "authlib==1.3.2",
    "httpx==0.28.0",
    "cryptography==43.0.3",
    "python-multipart==0.0.18",
    # PRD 2
    "boto3==1.35.0",
    "openai==1.57.0",
    "tiktoken>=0.7.0",
    "pypdf==5.1.0",
    "python-docx==1.1.2",
    "google-api-python-client==2.154.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.1",
]

[project.optional-dependencies]
test = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-mock==3.14.0",
    "httpx==0.28.0",
    "aiosqlite==0.20.0",
    "moto[sqs,s3]==5.0.0",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
